requirements.txt
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.38.0
gdown==5.2.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.25.1
imageio-ffmpeg==0.5.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
jiter==0.8.0
joblib==1.4.2
kiwisolver==1.4.4
matplotlib==3.7.0
moviepy==1.0.3
networkx==3.0
nltk==3.9.1
numpy==1.24.2
openai==1.57.0
packaging==23.0
Pillow==9.4.0
proglog==0.1.10
pybullet==3.2.6
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.1.3
Pygments==2.18.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
-e git+https://github.com/portal-cornell/robotouille.git@4190a1deabcd108320e6e4c694ba7fad55ea5b76#egg=robotouille
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tifffile==2023.2.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.14.0
